Blade View Directive Helpers Trait
This commit introduces the CompilesHelpers trait within the Hyperf\ViewEngine\Compiler\Concern namespace. The trait includes a method named compiled, which is responsible for compiling "d" statements into valid PHP code. The compiled method accepts a string of arguments and returns the corresponding PHP code to echo the "d" statement. This trait will be beneficial for enhancing the Blade view directive helpers in the Hyperf framework.